Module ransaq::db

Database querying and persistence.

ransaq stores crawled data in SQLite using sqlx.

Setup

To set up a development database you will need to install sqlx-cli,

cargo install sqlx-cli \
    --no-default-features \
    --features sqlite \
    --features rustls

create a .env file with a DATABASE_URL,

DATABASE_URL=sqlite:ransaq.sqlite

and finally run

sqlx database setup

Schema

You can view the schema by looking through the /migrations directory, or by using the sqlite1 command as follows:

sqlite3 -readonly ransaq.sqlite ".schema"

You will need to be running SQLite version 3.37.0 or later due to the use of STRICT tables (https://www.sqlite.org/releaselog/3_37_0.html) ↩

Re-exports

pub use glue::DbSerialize;

Modules

glue 🔒

Serialization logic necessary to put saq types into the database.

Macros

generate_upserts_by_name 🔒

Generates a method on Client named using the provided identifier which runs an upsert on the provided table name to make sure a row exists with the given name, returning the row’s id.

Structs

Client

Wraps all database logic.

ProductUpsertFields

Contains the necessary parameters to insert a row into the products table.

Functions

sqlite_configuration 🔒

Returns the SQLite configuration used by ransaq.

to_value_list 🔒

Encodes a list of IDs as a comma-separated string.