Macro ransaq::db::generate_upserts_by_name

macro_rules! generate_upserts_by_name {
    ($($fn:ident => $table:literal),*) => { ... };
}

Generates a method on Client named using the provided identifier which runs an upsert on the provided table name to make sure a row exists with the given name, returning the row’s id.