Struct ransaq::db::Client

pub struct Client {
    pool: SqlitePool,
}

Wraps all database logic.

Note - Client is both Sync and cheap to Clone thanks to SqlitePool being wrapped in an Arc.

Fields
pool: SqlitePool

The connection pool to use.

Implementations

impl Client

pub async fn new(url: &str) -> Result<Client>

Returns a new Client using the given url string.

See sqlite_configuration for accepted url formats.

pub async fn new_from_env() -> Result<Client>

Returns a new Client using the DATABASE_URL environment variable.

See sqlite_configuration for accepted formats.

pub async fn ensure_product_special_features(
&self,
product_id: i64,
special_feature_ids: Vec<i64>
) -> Result<()>

Uses upserts to make sure there are rows in product_special_features for each of the provided special_feature_ids.

updated at is always updated.

Any entries in product_special_features for the given product_id that don’t reference any of the provided special_feature_ids are subsequently deleted.

pub async fn ensure_product_grape_varieties(
&self,
product_id: i64,
grape_variety_ids_and_percentages: Vec<(i64, Option<u8>)>
) -> Result<()>

Uses upserts to make sure there are rows in product_grape_varieties for each of the provided (grape_variety_id, percentage) pairs.

updated_at and percentage are always updated to the provided values.

Any entries in product_grape_varieties for the given product_id that don’t reference the grape_variety_ids in the provided pairs are subsequently deleted.

pub async fn upsert_category(
&self,
name: &str,
url: &str,
parent_id: Option<i64>
) -> Result<i64>

Use an upsert query to make sure a row exists in the categories table with the provided name, and updating the url and parent_id fields.

Returns the row’s id.

pub async fn ensure_product_categories(
&self,
product_id: i64,
category_ids: Vec<i64>
) -> Result<()>

Uses upserts to make sure there are rows in product_categories for each of the provided category_ids.

updated_at is always updated.

Any entries in product_categories for the given product_id that don’t reference any of the provided category_ids are subsequently deleted.

impl Client

pub async fn upsert_product(
&self,
fields: ProductUpsertFields<'_>
) -> Result<i64>

Use an upsert query to ensure a row with the given saq_code exists in the products table, and update the remaining fields.

If a row already exists, updated_at will be set to the current time and consequently differ from created_at.

impl Client

pub async fn upsert_producer(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the producers table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_promoting_agent(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the promoting_agents table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_color(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the colors table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_region(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the regions table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_country(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the countries table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_grape_variety(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the grape_varieties table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_regulated_designation(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the regulated_designations table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_designation_of_origin(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the designations_of_origin table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_classification(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the classifications table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.

pub async fn upsert_special_feature(&self, name: &str) -> Result<i64>

Use an upsert to make sure there is a row in the special_features table with the given name.

Returns the row’s id.

Generated by the generate_upserts_by_name macro.