Struct ransaq::db::ProductUpsertFields

pub struct ProductUpsertFields<'a> {

    pub abv_percentage: Option<f32>,
    pub availability: &'a str,
    pub classification_id: Option<i64>,
    pub color_id: Option<i64>,
    pub container_count: Option<u8>,
    pub container_milliliters: Option<u32>,
    pub country_id: Option<i64>,
    pub description: &'a str,
    pub designation_of_origin_id: Option<i64>,
    pub image_url: &'a str,
    pub item_condition: &'a str,
    pub name: &'a str,
    pub price_cad: &'a f64,
    pub producer_id: Option<i64>,
    pub product_of_quebec: Option<&'a str>,
    pub promoting_agent_id: Option<i64>,
    pub region_id: Option<i64>,
    pub regulated_designation_id: Option<i64>,
    pub saq_code: &'a str,
    pub sugar_content_equality: Option<&'a str>,
    pub sugar_content_grams_per_liter: Option<f32>,
    pub upc_code: Option<&'a str>,

}

Contains the necessary parameters to insert a row into the products table.

Callers are responsible for using the necessary upsert_* and ensure_* methods on Client to obtain the necessary database ids to populate this struct. Note that these are all subject to FOREIGN KEY constraints.

More detail on the shape of this data can be found by reading through the saq docs.

This struct’s fields are kept in alphabetial order to make reasoning about which fields are included in queries easier.

Fields
abv_percentage: Option<f32>

The alcohol by volume percentage as a float.

availability: &'a str

The string representation of the ItemAvailability enum.

classification_id: Option<i64>

A database id from the classifications table.

color_id: Option<i64>

A database id from the colors table.

container_count: Option<u8>

The number of containers for the given product (i.e. 6 cans).

container_milliliters: Option<u32>

The number of milliliters contained in each container (i.e. 750ml).

country_id: Option<i64>

A database id from the countries table.

description: &'a str

The product’s description.

designation_of_origin_id: Option<i64>

A database id from the designations_of_origin table.

image_url: &'a str

A URL for an image of the product.

item_condition: &'a str

A string representation of the OfferItemCondition enum.

name: &'a str

The product’s name.

price_cad: &'a f64

The product’s price in Canadian Dollars as a float.

producer_id: Option<i64>

A database id from the producers table.

product_of_quebec: Option<&'a str>

A string representation of the ProductOfQuebec enum.

promoting_agent_id: Option<i64>

A database id from the promoting_agents table.

region_id: Option<i64>

A database id from the regions table.

regulated_designation_id: Option<i64>

A database id from the regulated_designations table.

saq_code: &'a str

The SAQ’s unique product identifier.

sugar_content_equality: Option<&'a str>

The string representation of the SugarContentEquality enum.

sugar_content_grams_per_liter: Option<f32>

The number of grams of sugar per liter as a float.

upc_code: Option<&'a str>

The product’s UPC code.