Module ransaq::saq::detailed_info

Parsing and cleanup logic to extract data out of the Detailed Info section of product pages.

Structs

DetailedInfo

Data extracted from the Detailed Info section of product pages.

GrapeVariety

The grape variety and percentage of it present in the product

Size

The product’s size

SugarContent

The product’s sugar content.

Enums

ProductOfQuebec

The specifics of the “Product of Québec” label

SugarContentEquality

Whether the actual sugar quantity is less than, greater than, or equal to grams_per_liter.

Functions

parse_abv

Converts a string indicating the alcohol by volume percentage into a float.

parse_grape_varieties

Converts the string representaiton of the grape varieties present in the product to a Vec of GrapeVariety.

parse_product_of_quebec

Converts the string representation of the “Product of Québec” label into the appropriate ProductOfQuebec enum variant.

parse_size

Converts the product’s size string (i.e. “6 x 200ml”) into a Size

parse_sugar_content

Converts the string representation of the product’s sugar content into a SugarContent.